HW02 InMemoryMealRepositoryImpl complete

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,11 +1,12 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.springframework.stereotype.Repository;
+import ru.javawebinar.topjava.AuthorizedUser;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -13,12 +14,13 @@ import java.util.concurrent.atomic.AtomicInteger;
  * GKislin
  * 15.09.2015.
  */
+@Repository
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.MEALS.forEach(this::save);
+        MealsUtil.MEALS.forEach(meal -> save(meal));
     }
 
     @Override
@@ -29,20 +31,36 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
         repository.put(meal.getId(), meal);
         return meal;
     }
-
+    // false if not found
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public boolean delete(int id, int userId) {
+        if (get(id, userId) != null) {
+            repository.remove(id);
+            return true;
+        }
+        else
+            return false;
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, int userId) {
+        Meal meal = repository.get(id);
+        if (userId == meal.getUserId())
+            return repository.get(id);
+        else
+            return null;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Collection<Meal> getAll(int userId) {
+        Collection<Meal> filteredColl = new ArrayList<>();
+        for (Meal meal : repository.values())
+        {
+            if (userId == meal.getUserId())
+                filteredColl.add(meal);
+        }
+        Collections.sort((List) filteredColl, (Comparator<Meal>) (o1, o2) -> o2.getDateTime().compareTo(o1.getDateTime()));
+        return filteredColl;
     }
 }
